IFS-2016-PF-SpendProfile-Action
Data Layer changes to use JPQL (Controller and Transactional service
changes)

+++ ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/competition/controller/CompetitionPostSubmissionController.java
@@ -84,7 +84,7 @@ public class CompetitionPostSubmissionController {
     }
 
     @GetMapping("/{competitionId}/count-pending-spend-profiles")
-    public RestResult<Integer> countPendingSpendProfiles(@PathVariable(value = "competitionId") Long competitionId) {
+    public RestResult<Long> countPendingSpendProfiles(@PathVariable(value = "competitionId") Long competitionId) {
         return competitionService.countPendingSpendProfiles(competitionId).toGetResponse();
     }
 }
