+++ ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/competition/transactional/CompetitionService.java
@@ -94,5 +94,5 @@ public interface CompetitionService {
 
     @PreAuthorize("hasAnyAuthority('project_finance')")
     @SecuredBySpring(value = "COUNT_PENDING_SPEND_PROFILES", description = "Project finance users can count projects for which Spend Profile generation is pending, for a given competition")
-    ServiceResult<Integer> countPendingSpendProfiles(Long competitionId);
+    ServiceResult<Long> countPendingSpendProfiles(Long competitionId);
 }
